Prints the first 98 fibonacci number

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -13,13 +13,13 @@ int main(void)
 	int i;
 	int f1 = 1;
 	int f2 = 2;
-	unsigned int f3;
+	long unsigned int f3;
 
 	printf("1, 2, ");
-	for(i = 0 ; i < 98 ; i++)
+	for (i = 0 ; i < 98 ; i++)
 	{
 		f3 = f1 + f2;
-		printf("%u, ", f3);
+		printf("%lu, ", f3);
 		f1 = f2;
 		f2 = f3;
 	}
